Move `NativeLibKind` from `rustc_session` to `rustc_hir`
